+++ doc/change_log.md
@@ -17,4 +17,4 @@
 * Material brush can now apply UV coordinates
 * Material brush can sample brush settings from under the cursor by pressing Shift-X
 * Adding face-vertices to ConvexBlockData
-
+* Creating vertex paint brush for painting face-vertices